feat(app): Installation page and refactoring
Renamed all DTOs to simply Request and Response classes. Created an endpoint that exposes info about the system letting the front-end know if the registration form (install) should be shown or not. Displaying this is controlled by middleware. Only one user can register openly and then the rest of the users should be created through the admin panel to close open registration.